refactor(config): move package graph and map logic to separate module
- Extracted package graph and map building functions from `main.rs` to a new `config.rs` module.
- Implemented `build_structures` function to create `PackageStructures` containing the graph and map.
- Moved topological sort, add_to_map, and add_to_graph functions to `config.rs`.
- Updated `main.rs` to use the new `config` module for building and sorting packages.
- Improved code organization and separation of concerns.
